+++ engine/records_database.go
@@ -14,7 +14,7 @@ func createDatabase() {
 
 func saveRecord(game *Game) {
 	db := database.Conn()
-	err := model.InsertRecord(db, game.User.Name, game.Score.Score)
+	err := model.InsertRecord(db, game.User.Name, game.Score.Score, int(game.Dificulty))
 	if err != nil {
 		log.Fatal(err)
 	}